refactor(donors): expose DonorPicker via feature barrel
donations and reports consumed DonorPicker through a deep internal path
(@/features/donors/donor-picker). Add src/features/donors/index.ts as
the feature's public API and route cross-feature imports through it, so
the module boundary is explicit rather than reaching into internals.
