+++ src/features/donors/index.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from 'vitest'
+import { DonorPicker } from './index'
+
+describe('donors public API', () => {
+  it('exposes DonorPicker', () => {
+    expect(DonorPicker).toBeTypeOf('function')
+  })
+})
