+++ src/features/donors/index.ts
@@ -1,0 +1,3 @@
+// Public API of the donors feature. Other features import from here, not from
+// deep internal paths.
+export { DonorPicker } from './donor-picker'
